backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
rawpy
numpy
tifffile
Pillow
imageio
requests
python-dotenv
imageio